+++ src/type_pool.cpp
@@ -98,6 +98,7 @@ TypePool::TypePool() {
   this->initErrorType(TYPE::SystemError, "SystemError");
   this->initErrorType(TYPE::StackOverflowError, "StackOverflowError");
   this->initErrorType(TYPE::RegexSyntaxError, "RegexSyntaxError");
+  this->initErrorType(TYPE::RegexMatchError, "RegexMatchError");
   this->initErrorType(TYPE::GlobbingError, "GlobbingError");
   this->initErrorType(TYPE::UnwrappingError, "UnwrappingError");
   this->initErrorType(TYPE::IllegalAccessError, "IllegalAccessError");