+++ CHANGELOG.md
@@ -98,6 +98,7 @@
     - use slice method instead
 - **Breaking Change**: ``Regex#replace`` method internally use ``pcre2_substitute`` api
     - now expand meta characters during replacement
+- **Breaking Change**: ``Regex`` methods throw ``RegexMatchError`` instead of ``InvalidOperationError``
 - now some builtin variables are constants
     - ``ON_ASSERT``, ``ON_ERR``, ``ON_EXIT``
     - ``TRUE``, ``True``, ``true``